backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9; python_version < "3.13"
psycopg[binary]==3.2.12; python_version >= "3.13"

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# PKI and cryptography
cryptography==41.0.7
hvac==2.0.0
docker==6.1.3
urllib3<2.0
requests<2.32

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Job scheduling for async tasks
apscheduler==3.10.4

# HTTP client for external calls
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Email notifications
aiosmtplib==3.0.1